requirements.txt
pyTelegramBotAPI
python-dotenv
moviepy
git+https://github.com/openai/whisper.git
